feat: implement auth (#11)
Implements auth using JWT. However:

Currently, this generates a JWT token from a secret (here called "token"). This is obviously not the end game. Soon™ we want to implement an identity provider based auth flow, i.e. the "token" would be a personal access token which is provided to the identity provider, which responds with a JWT token, which then is used to authenticate API requests.

That said, this is good for now / testing imo.